feat: move update thumbnail info modal to modal/ directory
