+++ src/components/modal/update-thumbnail-info/modal.scss
@@ -1,5 +1,5 @@
-@import "../../colors";
-@import "../../frameless";
+@import "../../../colors";
+@import "../../../frameless";
 
 .update-thumbnail-info-modal {
     margin-top: 30vh;
